requirements.txt
markdown
python-dotenv
python-git
pyspellchecker
textrazor